Three important things we learned
The first important thing we learned was Agile methods. Agile is an iterative approach to project management and software development that helps teams deliver value to their customers faster and with fewer headaches. Instead of betting everything on a "big bang" launch, an agile team delivers work in small, but consumable, increments. Our team produced user stories and started right away with a prototype for our customers while continuously refactoring the code to make the customer happy. We planned, design, developed, tested, deployed, and reviewed everything along the project if something did not work it was immediately addressed. Implementing the Agile method helped us deliver our final project without bugs.  

The second thing we learned in software development class was Python and the importance of testing, pytest. Pytest is a Python testing framework that originated from the PyPy project. It can be used to write several types of software tests, including unit tests, integration tests, end-to-end tests, and functional tests. Its features include parametrized testing, fixtures, and assert re-writing. This is important when writing code to be certain that the code and functions you have written work. Python is a high-level, general-purpose programming language. Its design philosophy emphasizes code readability with the use of significant indentation. Python is dynamically typed, and garbage collected. It supports multiple programming paradigms, including structured, object-oriented, and functional programming. We wrote all our projects in python, which was the language used most of the time in class. 

The third important thing I would say we learned is GitHub repository. Git is a distributed version control system: tracking changes in any set of files, usually used for coordinating work among programmers collaboratively developing source code during software development. Its goals include speed, data integrity, and support for distributed, non-linear workflows. We used GitHub to collaborate with other team members on the same project. We were able to track changes and keep up with what was happening with the project.